pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["beanslee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
graia-ariadne = {extras = ["standard"], version = "^0.11.2"}
graia-saya = "^0.0.17"
pillow = "^9.4.0"
requests = "^2.28.2"
graia-scheduler = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
